requirements.txt
# File: requirements.txt
#
# Daftar pustaka yang dibutuhkan.
# Instal dengan: pip install -r requirements.txt

tensorflow
streamlit
numpy
Pillow
opencv-python-headless
matplotlib
scipy
scikit-learn